Change encryption preference's summary conditionally
With this change Google added a summary to the encryption preference.

https://github.com/DirtyUnicorns/android_packages_apps_Settings/commit/415ce9a77d44732966fbab641d1ecaa8e8688381#diff-4b7e0e8b78bb0ce070384da41a76825f

The problem is that even if you're not encrypted, it says that your phone is.
Is very misleading to the user that don't bother to enter the actual preference
and want to see if they are. This fixes that.

Change-Id: Iaeaffd340f2e248748f244a947aa71eb8ad4ab94